Refactor db folder
