+++ src/database/queries/deliveryPartner.query.js
