+++ src/database/queries/otp.query.js
