+++ src/routes/restaurant.routes.js
@@ -16,8 +16,6 @@ router
   .patch(updateRestaurant)
   .delete(deleteRestaurant);
 
-// router.route("/recommended").get(getNearestRestaurants);
-
 router.route("/verify").post(verifyRestaurant);
 
 export default router;
